+++ src/components/shared/pagination/pagination.js
@@ -12,6 +12,8 @@ const Pagination = () => {
     const { currentPage, sortMode, doorsCount, pageSize } = useAppSelector(state => state.catalog);
     const pageCount = Math.ceil(doorsCount / pageSize);
     const pageNumbers = [];
+    const firstPage = 1;
+    const middlePage = Math.ceil(pageCount / 2);
 
     const scrollToTop = () => { // Вынести в хелперы.
         window.scrollTo({
@@ -29,7 +31,7 @@ const Pagination = () => {
         scrollToTop();
     }, [currentPage]);
 
-    for (let i = 1; i <= pageCount; i++) {
+    for (let i = 1; i <= pageCount; i++) { // Вынести в хелперы.
         pageNumbers.push(i);
     }
 
@@ -45,18 +47,71 @@ const Pagination = () => {
         dispatch(setCurrentPage(page));
     };
 
+    const getPageNumbers = () => {
+        if (pageCount <= 3) {
+            return (
+                pageNumbers.map(page => (
+                    <li 
+                        key={`page_${page}`}
+                        onClick={() => setPage(page)}
+                        className={page === currentPage && styles.active}
+                    >
+                        {page}
+                    </li>
+                ))
+            );
+        } else {
+            return (
+                <>
+                    <li 
+                        key={`page_${firstPage}`}
+                        onClick={() => setPage(firstPage)}
+                        className={currentPage === firstPage && styles.active}
+                    >
+                        {firstPage}
+                    </li>
+                    {currentPage !== firstPage && currentPage !== pageCount ? (
+                            <>
+                                <li>...</li>
+                                <li 
+                                    key={`page_${currentPage}`}
+                                    onClick={() => setPage(currentPage)}
+                                    className={currentPage === currentPage && styles.active}
+                                >
+                                    {currentPage}
+                                </li>
+                                <li>...</li>
+                            </>
+                        ) : (
+                            <>
+                                <li>...</li>
+                                    <li 
+                                        key={`page_${middlePage}`}
+                                        onClick={() => setPage(middlePage)}
+                                        className={middlePage === currentPage && styles.active}
+                                    >
+                                        {middlePage}
+                                    </li>
+                                <li>...</li>
+                            </>
+                        )   
+                    } 
+                    <li 
+                        key={`page_${pageCount}`}
+                        onClick={() => setPage(pageCount)}
+                        className={currentPage === pageCount && styles.active}
+                    >
+                        {pageCount}
+                    </li>
+                </>
+            );
+        }
+    };
+
     return (
         <ul className={styles.pagination}>
             {currentPage !== 1 && <li onClick={prevPage}>{'<'}</li>}
-            {pageNumbers.map(page => (
-                <li 
-                    key={`page_${page}`}
-                    onClick={() => setPage(page)}
-                    className={page === currentPage && styles.active}
-                >
-                    {page}
-                </li>
-            ))}
+            {getPageNumbers()}
             {currentPage !== pageCount && <li onClick={nextPage}>{'>'}</li>}
         </ul>
     );